DERDAT 1s duration option (#29)
* Adding_1s_input_option_shiny.R

Adding_input_duration_1s in shiny.R

* Flag_duration_change_in_criteria

change in criteria in "flag_duration_for_updating_if_value_non_standard_and_calced_interval_is_5s" function for the definition of flag duration.

* "Adding_1s_input_option_shiny.R"

1s input option is added in shiny.R

* remove NAs in UFLS_detection_voltage

NAs are removed during the mean voltage calculations.

* ignore_downsampling_for_1s_in_ideal_response

down-sampling of "ideal response" is ignored while analyzing 1s data because "ideal response" is already with 1s duration. Only column name "time_group" is changes to match with the existing data-frame.

* Suggested modification to "Ideal Response Curve"

Potential modification for "Ideal Response Curve" is suggested in "Assumptions and Methodology" tab of GUI

* Documenting minor discrepancies in the definition of "Reconnection Compliance Status"

Minor discrepancies in the definition of "Reconnection Compliance Status" is identified and mentioned in the documentation for future reference.

* downsampling "post event response" to 5s during 1s data analysis

"post event response" is down-sampled to 5s, only for 1s data analysis, for the definition of "reconnection status".

* Typo corrected in "Ideal Response function"

* Typo "duration" is corrected

* "d" duration is updated from the "reconnection data"

* spacing after "if" statement is corrected

* remove NAs during minimum and maximum f and v calculations

* adding if condition for downsampling from 1s to 5s

Down-sampling is done for data with 1s duration to 5s for "post_event_response"  and the other duration are allowed with the previous "post_event_response" calculations.

* fixed indenting

* downsampling from 1s to 5s for reconnection

Calculation of "post_event_response" is downsampled from 1s to 5s. Test file is created for dwonsampling.

* Spacing issue is fixed

Spacing in the added lines of codes in "island_assessment_function.R" and "ufls_detection_voltage.R" are resolved.

* Changes in documentation_tab regarding "reconnection_compliance_status"

* Spacing in "thicken" argument in "ideal_response_function.R"

* spacing inside if condition is fixed

* tidying unintended changes

* tidy up downsampling

---------

Co-authored-by: Phoebe <53634732+phoebeheywood@users.noreply.github.com>